add transferAuthIx to sdk, refactor tests
